tc: tl;dr -- fix for `super` keyword error.
```
yarn test:it
yarn run v1.22.5
$ testcafe chrome:headless ./test/it/*.spec.js
ERROR Cannot prepare tests due to an error.

/Users/tourdedave/_dev/eyes.sdk.javascript1/packages/eyes-sdk-core/lib/sdk/EyesClassic.js:311
        return super.checkWindowBase(
               ^^^^^

SyntaxError: 'super' keyword unexpected here
    at Object.<anonymous> (/Users/tourdedave/_dev/eyes.sdk.javascript1/packages/eyes-sdk-core/index.js:221:23)
    at Object.<anonymous> (/Users/tourdedave/_dev/eyes.sdk.javascript1/packages/eyes-testcafe/src/SpecDriver.js:1:35)
    at Object.<anonymous> (/Users/tourdedave/_dev/eyes.sdk.javascript1/packages/eyes-testcafe/test/it/SpecDriver.spec.js:2:14)
```

The `super` keyword is not supported on async class methods in the
latest stable version of Babel that is used by TestCafe. The workaround
is to call the method through prototypical inheritance.

re:
https://testcafe-discuss.devexpress.com/t/calling-super-class-async-method-from-extending-class/746
https://github.com/babel/babel/issues/3930

A proper fix for this will come out in Babel 7. So this will eventually
be something we can revert. Alternatively, there is a Babel plugin available
that applies this workaround automatically. If TestCafe adopts this
approach then we can revert this workaround sooner. Something to keep in
the back of our minds.