tc: revert previous commit and drop eyes-testcafe from the workspace
TestCafe processes code through Babel that ends up running in strict mode,
which causes `this` to not be passed. This breaks all of the `call` commands
in the previous commit. This presented once running an e2e coverage test.

By default TestCafe does not transpile code that lives in a local
node_modules directory. But when packages are used from the hoisted
node_modules directory in the yarn workspace, it attempts to process
them with Babel. There appears to be no way to configure this behavior.
The fix is to drop this package from the workspace and link packages
as needed, or rely on what's in the public registry.

This feels like a reasonable approach since we are actively considering
dropping yarn workspaces (e.g., PR 232).