tc: updated package.json scripts and deps
Specifically:
- dialed in running coverage tests in Docker on Chromium (to use
existing baselines)
- added explicit scripts for testing on different browsers (headless &
full - will circle back to see if there's a way to just use an env var
for this instead)
- updated to latest version of sdk-shared
- updated eyes-sdk-core to use `file:` to pick up changes for frame
traversal in EyesDriver. This was needed since yarn link won't work for
TestCafe b/c of the implicit Babeling of req'd sources outside of
node_modules, and, the Docker container won't follow the sym link (the
source needs to be in the current working directory).
- added the TestCafe Sauce plugin for future use