Page title lookup uses a Selector. When a title is not present in the AUT, the following error is presented:
```
Cannot obtain information about the node because the specified selector does not match any node in the DOM tree.
```

To compensate, the spec catches the error and returns an empty
string instead. Unfortunately, errors in TestCafe don't adhere to standard
calls like `error.message`, so scoping this behavior to just messages
like this is not possible.